backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
bs4==0.0.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==2.3.0
Django==5.2
django-browser-reload==1.18.0
django-environ==0.12.0
django-modeltranslation==0.19.16
django-nested-admin==4.1.1
django-tailwind==4.0.1
git-filter-repo==2.47.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
django-redis==5.4.0
